stack funtion fixed, updated, added, etc
