+++ inc/mem.h
@@ -8,7 +8,7 @@
 
 t_mt	*ft_mtzip(t_mt **lst);
 t_mt	*ft_mtunzip(void **data, size_t size);
-t_mt	*ft_mtnew(void *data);
+t_mt	*ft_mtnew(void *data, ...);
 t_mt	*ft_mtnew_chaos(void *data);
 void	ft_mtpush_data(t_mt **lst, void *data);
 void	ft_mtpush_data_back(t_mt **lst, void *data);
