+++ src/stack_utils/ft_stkadd_mtlist_back.c
@@ -1,31 +1,30 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_mtnew.c                                         :+:      :+:    :+:   */
+/*   ft_stkadd_mtlist_back.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/11 01:20:05 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/19 09:38:58 by danpalac         ###   ########.fr       */
+/*   Created: 2024/11/19 10:58:58 by danpalac          #+#    #+#             */
+/*   Updated: 2024/11/19 10:59:03 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mt.h"
+#include "stack.h"
 
-// Creates a new node for a list.
-t_mt	*ft_mtnew(void *data, ...)
+void	ft_stkadd_mtlist_back(t_stack **stack, t_mt **new)
 {
-	va_list	args;
-	t_mt	*new;
+	t_mt *tmp;
 
-	va_start(args, data);
-	new = (t_mt *)malloc(sizeof(t_mt));
-	if (!new)
-		return (NULL);
-    ft_memset(new, 0, sizeof(t_mt));
-	new->data = data;
-	new->size = va_arg(args, size_t);
-	new->count = va_arg(args, int);
-	va_end(args);
-	return (new);
+	if (!stack || !new)
+		return ;
+	tmp = *new;
+	while (tmp)
+	{
+		if (!(*stack))
+			(*stack) = ft_stknew(NULL, "temp");
+		ft_mtpush_back(&tmp, &(*stack)->head);
+	}
+	(*stack)->tail = ft_mtlast((*stack)->head);
+	(*stack)->size = ft_mtsize((*stack)->head);
 }
