+++ src/stack_utils/ft_stkmigrate_back.c
@@ -1,31 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_mtnew.c                                         :+:      :+:    :+:   */
+/*   ft_stkmigrate_back.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/11 01:20:05 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/19 09:38:58 by danpalac         ###   ########.fr       */
+/*   Created: 2024/11/19 10:57:27 by danpalac          #+#    #+#             */
+/*   Updated: 2024/11/19 10:57:32 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mt.h"
+#include "stack.h"
 
-// Creates a new node for a list.
-t_mt	*ft_mtnew(void *data, ...)
+void	ft_stkmigrate_back(t_stack **src, t_stack **dest)
 {
-	va_list	args;
-	t_mt	*new;
-
-	va_start(args, data);
-	new = (t_mt *)malloc(sizeof(t_mt));
-	if (!new)
-		return (NULL);
-    ft_memset(new, 0, sizeof(t_mt));
-	new->data = data;
-	new->size = va_arg(args, size_t);
-	new->count = va_arg(args, int);
-	va_end(args);
-	return (new);
+	if (!src || !(*src) || !(*src)->head)
+		return ;
+	ft_mtmigrate_back(&(*src)->head, &(*dest)->head);
+	(*src)->tail = ft_mtlast((*src)->head);
+	if ((*dest)->size == 0)
+		(*dest)->tail = (*dest)->head;
+	else
+		(*dest)->tail = ft_mtlast((*dest)->head);
+	(*src)->size = ft_mtsize((*src)->head);
+	(*dest)->size = ft_mtsize((*dest)->head);
 }
