+++ src/stack_utils/ft_stkprint.c
@@ -1,31 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_mtnew.c                                         :+:      :+:    :+:   */
+/*   ft_stkprint.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/11 01:20:05 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/19 09:38:58 by danpalac         ###   ########.fr       */
+/*   Created: 2024/11/19 11:00:22 by danpalac          #+#    #+#             */
+/*   Updated: 2024/11/19 11:00:28 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mt.h"
+#include "stack.h"
 
-// Creates a new node for a list.
-t_mt	*ft_mtnew(void *data, ...)
+void	ft_stkprint(t_stack *stack, int type, char *c)
 {
-	va_list	args;
-	t_mt	*new;
-
-	va_start(args, data);
-	new = (t_mt *)malloc(sizeof(t_mt));
-	if (!new)
-		return (NULL);
-    ft_memset(new, 0, sizeof(t_mt));
-	new->data = data;
-	new->size = va_arg(args, size_t);
-	new->count = va_arg(args, int);
-	va_end(args);
-	return (new);
+	if (!stack)
+	{
+		ft_printf("(Stack is NULL)\n");
+		return ;
+	}
+	ft_printf("Stack size: %d\n", stack->size);
+	ft_printf("Stack head: %p\n", stack->head);
+	ft_printf("Stack Name: %s\n", stack->name);
+	ft_mtprint(stack->head, type, c);
+	ft_printf("Stack tail: %p\n", stack->tail);
+	ft_printf("\n");
 }
