+++ src/stack_utils/ft_stkzip.c
@@ -1,31 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_mtnew.c                                         :+:      :+:    :+:   */
+/*   ft_stkzip.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/11 01:20:05 by danpalac          #+#    #+#             */
-/*   Updated: 2024/11/19 09:38:58 by danpalac         ###   ########.fr       */
+/*   Created: 2024/11/19 10:54:08 by danpalac          #+#    #+#             */
+/*   Updated: 2024/11/19 10:54:35 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "mt.h"
+#include "stack.h"
 
-// Creates a new node for a list.
-t_mt	*ft_mtnew(void *data, ...)
+t_stack	*ft_stkzip(t_stack **stack)
 {
-	va_list	args;
-	t_mt	*new;
+	t_stack	*zip;
+	t_mt	*zipdata;
 
-	va_start(args, data);
-	new = (t_mt *)malloc(sizeof(t_mt));
-	if (!new)
+	if (!(*stack) || !(*stack)->head)
 		return (NULL);
-    ft_memset(new, 0, sizeof(t_mt));
-	new->data = data;
-	new->size = va_arg(args, size_t);
-	new->count = va_arg(args, int);
-	va_end(args);
-	return (new);
+	zip = ft_stknew(NULL, "zip");
+	if (!zip)
+		return (NULL);
+	zipdata = ft_mtzip(&(*stack)->head);
+	ft_mtpush(&zipdata, &zip->head);
+	(*stack) = NULL;
+	return (zip);
 }
